M3-203 Feat : 픽셀 정보 헤더를 공통 위젯으로 분리
